+++ functions/NVDelSessionLists.php
@@ -7,6 +7,14 @@ function NVDelSessionLists($session, $listClass)
     // Ensure that the required global variables are set
     global $conn, $appid, $uuid;
 
+    if (!isset($conn, $appid, $uuid)) {
+        error_log("NVDelSessionLists: Required variables are not set.");
+        return false;
+    }
+
+    // Not allowed when called from front-end without being sanitized
+    if (AFIsUnsafe()) return false;
+
     $stmt = $conn->prepare("DELETE FROM List WHERE AppID = ? AND UserID = ? AND SessionID = ? AND Class = ?");
     $stmt->bind_param("ssss", $appid, $uuid, $session, $listClass);
